Finally fixed issue, where data read from EditScript was not properly aligned for JNA Structures. Added ExternalScannerState and TSPoint to solve misalignment; Added remaining unit tests for EditScripts to see if JNA translation works properly;

+++ treesitter-truediff-API/src/main/java/inca/treesitterAPI/editscriptAPI/Attach.java
@@ -10,11 +10,12 @@ public class Attach extends Structure {
     public short tag;
     public int link;
     public Pointer parent_id;
-    public short parent_tag;
+    public int parent_tag;
 
     public Attach() {
         super();
     }
+
     public Attach(Pointer p) {
         super(p);
         this.read();
